+++ python/vtext/tokenize.py
@@ -5,7 +5,7 @@
 # modified, or distributed except according to those terms.
 
 from ._lib import BaseTokenizer
-from ._lib import UnicodeSegmentTokenizer
+from ._lib import UnicodeWordTokenizer
 from ._lib import RegexpTokenizer
 from ._lib import VTextTokenizer
 from ._lib import CharacterTokenizer
@@ -13,7 +13,7 @@ from ._lib import CharacterTokenizer
 
 __all__ = [
     "BaseTokenizer",
-    "UnicodeSegmentTokenizer",
+    "UnicodeWordTokenizer",
     "RegexpTokenizer",
     "VTextTokenizer",
     "CharacterTokenizer",